+++ apps/fabric-website/src/pages/Controls/ControlsAreaPage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Page, IPageProps, withPlatform } from '@uifabric/example-app-base/lib/index2';
+import { Page, IPageProps, PlatformContext } from '@uifabric/example-app-base/lib/index2';
 import { getSubTitle } from '../../utilities/index';
 import { Platforms } from '../../interfaces/Platforms';
 import { IPageJson } from 'office-ui-fabric-react/lib/common/DocPage.types';
@@ -22,4 +22,6 @@ const ControlsAreaPageBase: React.StatelessComponent<IControlsPageProps> = props
   return <Page subTitle={getSubTitle(props.platform)} jsonDocs={jsonDocs} {...props} />;
 };
 
-export const ControlsAreaPage: React.StatelessComponent<IControlsPageProps> = withPlatform<Platforms>(ControlsAreaPageBase);
+export const ControlsAreaPage: React.StatelessComponent<IPageProps<Platforms>> = (props: IPageProps<Platforms>) => (
+  <PlatformContext.Consumer>{(platform: Platforms) => <ControlsAreaPageBase platform={platform} {...props} />}</PlatformContext.Consumer>
+);