+++ apps/fabric-website/src/pages/Styles/Colors/ProductsPage.tsx
@@ -1,5 +1,12 @@
 import * as React from 'react';
-import { Markdown, ColorPalette, IColor, IColorPaletteTheme, IPageSectionProps } from '@uifabric/example-app-base/lib/index2';
+import {
+  Markdown,
+  MarkdownHeader,
+  ColorPalette,
+  IColorSwatch,
+  IColorPaletteTheme,
+  IPageSectionProps
+} from '@uifabric/example-app-base/lib/index2';
 import { IStylesPageProps, StylesAreaPage } from '../StylesAreaPage';
 import { ColorsProductsPageProps } from './ProductsPage.doc';
 
@@ -57,7 +64,7 @@ export class ColorsProductsPage extends React.Component<IStylesPageProps, IColor
                 <ColorPalette colors={AppColorSwatches} onColorSelected={this._changeApp} />
                 {activeAppColorPalette && (
                   <>
-                    <h2>{activeAppColorPalette.name}</h2>
+                    <MarkdownHeader as="h2">{activeAppColorPalette.name}</MarkdownHeader>
                     <p>{activeAppColorPalette.notes}</p>
                     <ColorPalette colors={activeAppColorPalette.colors} />
                   </>
@@ -87,7 +94,7 @@ export class ColorsProductsPage extends React.Component<IStylesPageProps, IColor
     }
   };
 
-  private _changeApp = (color: IColor) => {
+  private _changeApp = (color: IColorSwatch) => {
     const { activeAppColorPalette } = this.state;
 
     // Show the active app's color palette (optional).