+++ packages/example-app-base/src/components/Page/sections/ExamplesSection.tsx
@@ -1,19 +1,16 @@
 import * as React from 'react';
 import { ExampleCard } from '../../ExampleCard/index';
-import { pascalize } from '../../../utilities/index2';
-import { IExample, IPageSectionProps } from '../Page.types';
+import { IExample, IPageSectionPropsWithSectionName } from '../Page.types';
 import * as styles from '../Page.module.scss';
 import { ICodeSnippetProps } from '../../CodeSnippet/index';
 
-export interface IExamplesSectionProps extends IPageSectionProps {
+export interface IExamplesSectionProps extends IPageSectionPropsWithSectionName {
   exampleKnobs?: React.ReactNode;
   examples: IExample[];
 }
 
 export const ExamplesSection: React.StatelessComponent<IExamplesSectionProps> = props => {
-  const { className, examples, exampleKnobs, sectionName = 'Usage', style } = props;
-  const { readableSectionName = sectionName } = props;
-  const sectionId = pascalize(sectionName);
+  const { className, examples, exampleKnobs, sectionName, readableSectionName, style, id } = props;
   const codeHighlighterProps: ICodeSnippetProps = {
     language: 'typescript'
   };
@@ -21,8 +18,8 @@ export const ExamplesSection: React.StatelessComponent<IExamplesSectionProps> = 
   return (
     <div className={className} style={style}>
       <div className={styles.sectionHeader}>
-        <h2 className={styles.subHeading} id={sectionId}>
-          {readableSectionName}
+        <h2 className={styles.subHeading} id={id}>
+          {readableSectionName || sectionName}
         </h2>
       </div>
       <div>