+++ packages/experiments/src/components/Button/SplitButton/SplitButton.tsx
@@ -3,14 +3,13 @@ import { createComponent } from '../../../Foundation';
 import { useSplitButtonState as state } from './SplitButton.state';
 import { SplitButtonStyles as styles, SplitButtonTokens as tokens } from './SplitButton.styles';
 import { ISplitButtonProps } from './SplitButton.types';
-import { SplitButtonView as view } from './SplitButton.view';
+import { SplitButtonView } from './SplitButton.view';
 
-export const SplitButton: React.StatelessComponent<ISplitButtonProps> = createComponent({
+export const SplitButton: React.StatelessComponent<ISplitButtonProps> = createComponent(SplitButtonView, {
   displayName: 'SplitButton',
   state,
   styles,
-  tokens,
-  view
+  tokens
 });
 
 export default SplitButton;