+++ packages/experiments/src/components/Button/SplitButton/SplitButton.types.tsx
@@ -77,6 +77,7 @@ export interface ISplitButtonProps
       | 'checked'
       | 'allowDisabledFocus'
       | 'ariaLabel'
+      | 'keytipProps'
       | 'defaultExpanded'
       | 'expanded'
       | 'onKeyDown'