+++ packages/experiments/src/components/Button/SplitButton/SplitButton.view.tsx
@@ -17,6 +17,7 @@ export const SplitButtonView: ISplitButtonComponent['view'] = props => {
     onClick,
     allowDisabledFocus,
     ariaLabel,
+    keytipProps,
     expanded,
     root,
     button,
@@ -69,6 +70,7 @@ export const SplitButtonView: ISplitButtonComponent['view'] = props => {
         allowDisabledFocus={allowDisabledFocus}
         ariaLabel={menuButtonAriaLabel}
         onClick={onSecondaryActionClick}
+        keytipProps={keytipProps}
         menu={menu}
         componentRef={menuButtonRef}
       />