+++ packages/fluent-theme/src/fluent/styles/SearchBox.styles.ts
@@ -1,0 +1,15 @@
+import { ISearchBoxStyleProps, ISearchBoxStyles } from 'office-ui-fabric-react/lib/SearchBox';
+
+export const SearchBoxStyles = (props: ISearchBoxStyleProps): Partial<ISearchBoxStyles> => {
+  const { theme } = props;
+  const { effects, palette } = theme;
+
+  return {
+    root: [
+      {
+        borderRadius: effects.roundedCorner2,
+        borderColor: palette.neutralSecondaryAlt
+      }
+    ]
+  };
+};