+++ packages/office-ui-fabric-react/src/components/Panel/Panel.styles.ts
@@ -13,6 +13,8 @@ import {
   IStyle
 } from '../../Styling';
 import { FontWeights } from '../../Styling';
+import { getWindow } from '../../Utilities';
+
 // TODO -Issue #5689: Comment in once Button is converted to mergeStyles
 // import { IStyleFunctionOrObject } from '../../Utilities';
 // import { IButtonStyles, IButtonStyleProps } from '../../Button';
@@ -202,7 +204,8 @@ export const getStyles = (props: IPanelStyleProps): IPanelStyles => {
   const { palette, effects } = theme;
   const classNames = getGlobalClassNames(GlobalClassNames, theme);
   const isCustomPanel = type === PanelType.custom || type === PanelType.customNear;
-  const windowHeight = typeof window !== 'undefined' ? window.innerHeight : '100%';
+  const win = getWindow();
+  const windowHeight = typeof win !== 'undefined' ? win.innerHeight : '100%';
 
   return {
     root: [