+++ packages/react-cards/src/components/Card/CardItem/CardItem.ts
@@ -2,13 +2,12 @@ import * as React from 'react';
 import { createComponent } from '@uifabric/foundation';
 import { CardItemStyles as styles, CardItemTokens as tokens } from './CardItem.styles';
 import { ICardItemProps } from './CardItem.types';
-import { CardItemView as view } from './CardItem.view';
+import { CardItemView } from './CardItem.view';
 
-export const CardItem: React.FunctionComponent<ICardItemProps> = createComponent({
+export const CardItem: React.FunctionComponent<ICardItemProps> = createComponent(CardItemView, {
   displayName: 'CardItem',
   styles,
-  tokens,
-  view
+  tokens
 });
 
 export default CardItem;