+++ packages/fluent-theme/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/fluent-theme
 
-This log was last generated on Tue, 30 Apr 2019 12:31:36 GMT and should not be manually modified.
+This log was last generated on Tue, 14 May 2019 07:50:30 GMT and should not be manually modified.
+
+## 0.16.9
+Tue, 14 May 2019 07:50:30 GMT
+
+### Patches
+
+- Update Fabric assets link
+
+## 0.16.8
+Tue, 14 May 2019 01:56:43 GMT
+
+### Patches
+
+- Update border colors during rest state for certain components
 
 ## 0.16.7
 Tue, 30 Apr 2019 12:31:36 GMT